control_plane: Back out of running the script from the control plane
Since at the moment we're lacking any synchronization between the
control plane and the worker threads, we cannot guarantee that the
main thread will run through the script (and set all hooks) before
the worker threads being their routines. So for the time being
undo the changes to the control plane until things in thread module
get re-arranged.