serialize: Remove broken support for dumping functions not at top of the stack
dump_function_bytecode() allows the called specify at which index the
function to dump is located. But lua_dump() expects the function to
always be at the top of the stack, so the behaviour is broken.

Since we currently don't have a use case for dumping functions that
aren't at the top the stack, just make the dump_function_bytecode()
more limited.