serialize: Keep track and allow fetching of values marked for collection
If a value is detected to be a "special" wrapper that marks the value
for collection, i.e. a single entry table with 'collect' meta-field
set, then remember this property when serializing and later
deserializing the wrapped value.

Also, allow the users to fetch the deserialized values that were
marked for collection by specifying the wrapper object ID (i.e. its
address).