script: Support for collecting (copying) locals from worker threads
Mark local variables for collection assigning to them reference a
"special" value wrapper object. The wrapped values will be unwrapped
when transferred to hooks.

When worker threads have finished, values marked for collection will
be copied back to the main state and collected in a table that
replaces the wrapper object.