tcp_rr: Get initial epoll events for client socket from options
The behavior stays the same since in request-response scenario the
server waits for the client to send the first request, so we will
never have anything to read from the socket.

The benefit being that this unifies the run_client routine in both
tcp workloads, modulo the helper for running the epoll loop that they
run.