tests: Make exec-order test immune to client/server process startup order
Error messages from a client process that cannot connect to the
server because the other process has not started yet sometimes break
the test because the expected output on stderr doesn't match.

Switch to using stdout and filter out only the output lines we care
about to get rid of spurious failures.