workload: Create a full fledged flow for listening socket
While it does not apply to TCP workloads, for UDP workloads the
"listening" socket is also the receiving socket so we need to create
a flow for it sample the number of received bytes at regular
intervals.

The flow tear down is not as clean as it could be. We needlessly close
the socket twice.