workload: Multiplex server port based on run-time options
Introduce a run-time option for enabling server port multiplexing.

It is on by default for all workloads to keep the workload behavior
same as it was. It currently can't be turned off due to a limitation
of the command line flag parser.

This change is effectively has no effect. It prepares ground for
making server port multiplexing optional.